asylum: Major rearrangement of the site
Pretty much noone ever used the menu, so lets drop that, and move its
contents elsewhere: into a small button bar on the top right, and a
link farm at the bottom.

With this, the main article has considerably less wasted space around
it, and all important information is readily visible, without having
to navigate a menu.

The "recent posts" area is also turned off, as recent posts are now
expanded in the bottom link farm.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>